Refactor login to use MVP
