+++ app/src/main/java/org/wit/hillfort/activities/LandingActivity.kt
@@ -6,6 +6,7 @@ import android.view.Menu
 import kotlinx.android.synthetic.main.activity_landing.*
 import org.jetbrains.anko.intentFor
 import org.wit.hillfort.R
+import org.wit.hillfort.views.login.LoginView
 import org.wit.hillfort.views.signup.SignupView
 
 class LandingActivity: AppCompatActivity() {
@@ -21,7 +22,7 @@ class LandingActivity: AppCompatActivity() {
 
     // Start login activity when button is pressed
     btnLogin.setOnClickListener{
-      startActivityForResult(intentFor<LoginActivity>(), 0)
+      startActivityForResult(intentFor<LoginView>(), 0)
     }
   }
